+++ docs/sphinx/changelog.rst
@@ -14,3 +14,4 @@ New Features
 - Added linear elasticity for the internal particle (:merge:`1`). By `Nathan Miller`_.
 - Added the calculation of the current distance between two points in the particles (:merge:`2`). By `Nathan Miller`_.
 - Added the calculation of the decomposition of a vector into normal and tangential components w.r.t. another vector (:merge:`2`). By `Nathan Miller`_.
+- Added the calculation of the linear traction separation energy (:merge:`3`). By `Nathan Miller`_.